FEAT: Frontend now gets the query from the API
