+++ streamlit/Dockerfile
@@ -3,12 +3,6 @@ FROM python:3.9-slim
 WORKDIR /streamlit/app
 COPY requirements.txt /streamlit/app/requirements.txt
 
-RUN apt-get update && apt-get install -y \
-    build-essential \
-    curl \
-    software-properties-common \
-    git \
-    && rm -rf /var/lib/apt/lists/*
 
 RUN apt-get update \
     && apt-get -y install libpq-dev gcc --fix-missing
